fix: 예외 처리 수정
